# PERF: optimize is_numeric_v_string_like