@jbrockmendel 2021-03-18
nice!  two questions, neither blockers for this PR

1) IIUC this check exists to prevent a numpy-issued deprecation/future warning.  Do we know when that ceases to be an issue?

2)  Looks like is_string_like_dtype is only used in one other place (in dtypes.missing).  lets just rip it out