@jbrockmendel 2021-03-18
> Or I could update the helper function to what I used here, and then keep using it in those 3 places?

either way i guess.  i prefer to inline the check since i frequently find myself having to go look up what the is_foo_dtype's exact behavior is