@jorisvandenbossche 2021-03-18
OK, I am certainly fine with inlining here